Add comments to new question frontend, clean code
