+++ frontend/src/App.tsx
@@ -1,14 +1,14 @@
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import NewQuestion from "./pages/NewQuestion";
 
 function App() {
   return (
-    <Router>
+    <BrowserRouter>
       <Routes>
         <Route path="/question" element={<>question page list</>} />
         <Route path="/question/new" element={<NewQuestion />} />
       </Routes>
-    </Router>
+    </BrowserRouter>
   );
 }
