+++ frontend/src/components/NewQuestion/QuestionImage.tsx
