+++ frontend/src/components/NewQuestion/QuestionImageContainer.tsx
