+++ frontend/src/components/NewQuestion/QuestionImageDialog.tsx
