+++ frontend/src/main.tsx
@@ -1,33 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#8FB8ED",
-      contrastText: "#FFFFFF",
-    },
-    secondary: {
-      main: "#ECECEC",
-      contrastText: "#757575",
-    },
-  },
-  components: {
-    MuiButton: {
-      styleOverrides: {
-        root: {
-          textTransform: "none",
-          fontWeight: "bold",
-          // ":hover": {
-          //   backgroundColor: "",
-          // },
-        },
-      },
-    },
-  },
-});
+import theme from "./theme";
+import { ThemeProvider } from "@mui/material/styles";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
